requirements.txt
Django==5.0
django_environ==0.11.2
django_import_export==3.3.4
environ==1.0
